requirements.txt
tensorflow
keras
pandas
numpy
pyais
geopy
scikit-learn
scipy
sklearn
tqdm
regex
kafka-python
folium
matplotlib